requirements.txt
Django==3.2.6
djangorestframework==3.12.4
mysqlclient==2.0.3
django-cors-headers==3.7.0
jdatetime==3.6.2
persiantools==2.1.2
elasticsearch==7.14.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.1
django-redis==5.0.0
django-debug-toolbar==3.2.1
